frontend/server/requirements-testing.txt
coverage
mongomock
nose
nose-watch
pep8
pylint
pylint-quotes
requests_mock
